refactor: shared/ui > 세분화
