+++ src/shared/lib/types.ts
@@ -1,0 +1,5 @@
+import { PropsWithChildren, ComponentPropsWithoutRef, ElementType } from 'react';
+
+export type WithClassName<T extends ElementType> = PropsWithChildren<ComponentPropsWithoutRef<T>> & {
+  className?: string;
+};
